improve examples organization
