+++ src-lua/example/demo/asset.png
