+++ src-lua/example/test_auto/conf.lua
@@ -1,0 +1,6 @@
+function love.conf(t)
+  t.window.title = "Feather Auto Test"
+  t.window.width = 800
+  t.window.height = 600
+  t.console = false
+end
