+++ src-lua/example/test_cli/conf.lua
@@ -1,6 +1,6 @@
 function love.conf(t)
-  t.window.title  = "test_cli — no feather"
+  t.window.title = "Feather CLI Example"
   t.window.width  = 640
   t.window.height = 400
-  t.console       = false
+  t.console = false
 end
